EM suoristaminen, keskittäminen, solmujen pohjien suoristaminen

+++ core/src/fi/ymcafinland/demo/logiikka/Solmu.java
@@ -81,7 +81,7 @@ public class Solmu {
     }
     public void setMutsi(Solmu solmu){mutsi = solmu;}
 
-    public void setSijainti(int x, int y) {
+    public void setSijainti(float x, float y) {
         this.x = x;
         this.y = y;
     }
